[FLINK-23126] Refactor smoke-e2e into smoke-e2e-common and smoke-e2e-embedded
This closes #242.
