+++ statefun-e2e-tests/pom.xml
@@ -36,7 +36,8 @@ under the License.
         <module>statefun-e2e-tests-common</module>
         <module>statefun-sanity-e2e</module>
         <module>statefun-exactly-once-remote-e2e</module>
-        <module>statefun-smoke-e2e</module>
+        <module>statefun-smoke-e2e-common</module>
+        <module>statefun-smoke-e2e-embedded</module>
     </modules>
 
     <build>
