+++ statefun-e2e-tests/statefun-smoke-e2e-common/src/main/protobuf/internal.proto
