+++ statefun-e2e-tests/statefun-smoke-e2e-embedded/src/test/java/org/apache/flink/statefun/e2e/smoke/SmokeVerificationEmbeddedE2E.java
@@ -1,0 +1,34 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.statefun.e2e.smoke;
+
+import org.junit.Test;
+
+public class SmokeVerificationEmbeddedE2E {
+
+  @Test(timeout = 1_000 * 60 * 10)
+  public void runWith() throws Throwable {
+    ModuleParameters parameters = new ModuleParameters();
+    parameters.setNumberOfFunctionInstances(128);
+    parameters.setMessageCount(100_000);
+    parameters.setMaxFailures(1);
+
+    SmokeRunner.run(parameters);
+  }
+}
