+++ statefun-e2e-tests/statefun-smoke-e2e-embedded/src/test/resources/log4j.properties
